@@ -8,7 +8,7 @@ exclude = ["third_party"]
 
 [tool.ruff]
 line-length = 127
-target-version = "py310"
+target-version = "py311"
 exclude = ["third_party"]
 
 [tool.ruff.isort]
@@ -25,7 +25,8 @@ testpaths = ["tests"]  # ignore third_party dir for now
 exclude = [
     "learned_planner/notebooks/create_hard_reward.py",
     "learned_planner/notebooks/plot_planning.py",
-    "plot/behavior_analysis.py",
+    "plot/**",
+    "alternative-levels/**",
     "wandb/**",  # Saved old codes
     "third_party/**",  # Other libraries
     "build/**", # build directory for strict editable install
@@ -59,14 +60,14 @@ exclude_lines = [
 include = ["learned_planner*", "experiments"]
 
 [tool.setuptools.package-data]
-"*" = ["../.circleci/*.yml", "../k8s/*.yaml"]
+"*" = ["../.circleci/*.yml", "../k8s/*.yaml", "../plot/*.svg"]
 
 [project]
 name = "learned-planner"
 version = "0.1.0"
 description = "Planning behavior in RNNs that plays Sokoban."
 readme = "README.md"
-requires-python = ">=3.10"
+requires-python = ">=3.11"
 license = {file = "LICENSE.txt"}
 keywords = []
 
@@ -80,30 +81,31 @@ maintainers = []
 classifiers = []
 
 dependencies = [
-    "accelerate ~=0.32.1",
-    "matplotlib ==3.9.0",
     "gymnasium ~= 0.29.0",
     "wandb ~=0.17.4",
     "moviepy ~=1.0.3",
     "shimmy ~=1.1.0",  # match version in dev
     "optree ~= 0.10.0",
+    "plotly ~=5.20.0",
     # Freeze the numpy version so OpenBLAS numpy gets installed with pip. Conda-default MKL numpy causes errors when
     # running SubprocessVecEnv on the EPYC machines of Flamingo.
     "numpy ==1.26.4",
     "scikit-learn ~=1.4.0",
     "cairosvg ~=2.7.1",
+    "transformer-lens ~=2.4.0",
     "names-generator ~=0.2.0",
     "torchvision ~=0.18.1",
     "stable_baselines3 @ git+https://github.com/AlignmentResearch/learned-planners-stable-baselines3.git",
     "gym_sokoban @ git+https://github.com/AlignmentResearch/gym-sokoban.git",
     "farconf @ git+https://github.com/AlignmentResearch/farconf.git",
     "lp_cleanba @ git+https://github.com/AlignmentResearch/lp-cleanba.git@ef059c3d7e1d22515cc9d18e5927dc49db60cf1c",
-    "mamba_lens @ git+https://github.com/Phylliida/MambaLens.git",    
+    "mamba_lens @ git+https://github.com/taufeeque9/MambaLens.git",
 ]
 
 [project.optional-dependencies]
 torch = [
-    "torch ==2.3.1",
+    "torch ==2.4.0",
+    "torchvision ==0.19",
 ]
 dev = [
     # local development
@@ -122,6 +124,10 @@ dev = [
     "pytest ~=8.1.1",
     "pytest-cov >=4,<5",
     "pytest-env ~=1.1.2",
+
+    # launch jobs
+    "names_generator ~=0.1.0",
+    "GitPython ~=3.1.41",
 ]
 
 [project.urls]
