@@ -18,13 +18,13 @@ def train_probe(
         base_save_prefix=training_mount,
         cmd=TrainProbeConfig(
             dataset_path=dataset_path,
-            policy_path=training_mount / "c51papdh",
+            policy_path="drc33/bkynosqi/cp_2002944000",
             learning_rate=1e-4,
             epochs=5000,
             eval_epoch_interval=500,
             batch_size=512,
             weight_decay=1e-1,
-            weight_decay_norm=1,
+            weight_decay_type="l1",
             eval_type="probe",
             device=device,
             seed=seed,
@@ -35,8 +35,8 @@ def train_probe(
 def eval_probe(
     training_mount: Path = DEFAULT_TRAINING,
     eval_type: str = "probe",
-    policy_path: Path = DEFAULT_TRAINING / "c51papdh",
-    eval_model: Optional[str] = None,
+    policy_path: str = "drc33/bkynosqi/cp_2002944000",
+    probe_path: Optional[str] = None,
     seed: int = random_seed(),
 ):
     return WandbCommandConfig(
@@ -45,7 +45,7 @@ def eval_probe(
             policy_path=policy_path,
             eval_only=True,
             eval_type=eval_type,
-            eval_model=eval_model,
+            probe_path=probe_path,
             seed=seed,
         ),
     )
@@ -53,3 +53,4 @@ def eval_probe(
 
 # fmt: off
 def train_local(): return train_probe("cpu", Path("."))
+def train_cluster(): return train_probe("cpu", DEFAULT_TRAINING)
