@@ -1,4 +1,4 @@
-import gym_sokoban  # noqa
+import gym_sokoban  # noqa: F401 # pyright: ignore
 import gymnasium as gym
 
 
