@@ -2,24 +2,10 @@
 import pathlib
 import re
 
-import matplotlib
 import matplotlib.pyplot as plt
 import pandas as pd
 
-style = {
-    "font.family": "serif",
-    "font.serif": "Times New Roman",
-    "mathtext.fontset": "cm",
-    "font.size": 10,
-    "legend.fontsize": 10,
-    "axes.titlesize": 10,
-    "axes.labelsize": 10,
-    "xtick.labelsize": 10,
-    "ytick.labelsize": 10,
-    "figure.figsize": (3.25, 2),
-    "figure.constrained_layout.use": True,
-}
-matplotlib.rcParams.update(style)
+import learned_planner.interp.plot  # noqa
 
 steps_to_think_for_pe = [0, 2, 4, 8, 12, 16]
 network_name = "drc33"
@@ -41,7 +27,7 @@ df = df[steps_to_think_for_pe]
 csv_file = pathlib.Path(__file__).parent / "data" / f"resnet_{dataset_name}_success_across_thinking_steps.csv"
 df_resnet = pd.read_csv(csv_file, index_col="Step")
 
-per_step = df
+per_step = df * 100
 # per_step = per_step - per_step.loc[0]
 
 fig, axes = plt.subplots(2, 1, figsize=(3.25, 2.5), sharex=True, height_ratios=[3, 1])
@@ -51,9 +37,8 @@ for i in range(len(per_step.T)):
     per_step.iloc[:, i].plot(color=plt.get_cmap("viridis")(this_step_proportion), ax=axes[0], legend=False)
 
 resnet_run_name = str(df_resnet.columns[0]).split(" - ")[0]
-df_resnet[f"{resnet_run_name} - {dataset_name}/00_episode_successes"].plot(color="C1", ax=axes[0], label="resnet")
+(df_resnet[f"{resnet_run_name} - {dataset_name}/00_episode_successes"] * 100).plot(color="C1", ax=axes[0], label="resnet")
 (per_step.max(axis=1) - per_step[0]).plot(ax=axes[1], label="valid_medium", color="C0")
-y_label = dict(test_unfiltered="Test-unfiltered", valid_medium="Val-medium", hard="Hard")[dataset_name]
 
 dataset_name = "hard"
 csv_file = pathlib.Path(__file__).parent / "data" / f"{network_name}_{dataset_name}_success_across_thinking_steps.csv"
@@ -70,7 +55,7 @@ df = df[steps_to_think_for_pe]
 csv_file = pathlib.Path(__file__).parent / "data" / f"resnet_{dataset_name}_success_across_thinking_steps.csv"
 df_resnet = pd.read_csv(csv_file, index_col="Step")
 
-per_step = df
+per_step = df * 100
 (per_step.max(axis=1) - per_step[0]).plot(ax=axes[1], label="hard", color="C3")
 
 axes[0].grid(True)
@@ -78,11 +63,15 @@ axes[1].grid(True)
 
 axes[1].set_xlabel("Environment steps, training")
 
-axes[0].set_ylabel(y_label + " solved")
-axes[1].set_ylabel("Plan. Effect")
+axes[0].set_ylabel(r"% solved")
+axes[1].set_ylabel("Planning Effect")
 axes[0].legend(ncols=3, prop={"size": 8})
 axes[0].set_xlim((998400.0, 2002944000.0))
-# axes[1].legend(ncols=2, prop={"size": 8})
-plt.savefig(plots_dir / "mixed_fig1_curve.pdf", format="pdf")
+# set xticks
+axes[0].set_xticks([7e7, 5e8, 1e9, 1.5e9, 2e9])
+# set xticklabels
+axes[0].set_xticklabels(["70M", "500M", "1B", "1.5B", "2B"])
+axes[1].legend(ncols=2, prop={"size": 8})
+plt.savefig(plots_dir / "fig1_with_valid_hard_plan_effect.pdf", format="pdf")
 plt.show()
 plt.close()
